requirements.txt
flask
flask-cors
torch
torchvision
pillow
opencv-python
matplotlib
numpy